+++ packages/visual-editor/locales/platform/cs/visual-editor.json
@@ -523,7 +523,6 @@
     "search": "Hledat písma ..."
   },
   "fontSize": "Velikost písma",
-  "fontSizeDefaultLabel": "Výchozí",
   "friday": "pátek",
   "getDirections": "Získejte pokyny",
   "getDirectionsForLocation": "Získat trasu pro {{address}}",
