+++ packages/visual-editor/locales/platform/da/visual-editor.json
@@ -523,7 +523,6 @@
     "search": "Søg skrifttyper ..."
   },
   "fontSize": "Skrifttypestørrelse",
-  "fontSizeDefaultLabel": "Misligholdelse",
   "friday": "fredag",
   "getDirections": "Få rutevejledning",
   "getDirectionsForLocation": "Få rutevejledning til {{address}}",
