+++ packages/visual-editor/locales/platform/de/visual-editor.json
@@ -523,7 +523,6 @@
     "search": "Schriftarten suchen ..."
   },
   "fontSize": "Schriftgröße",
-  "fontSizeDefaultLabel": "Standard",
   "friday": "Freitag",
   "getDirections": "Route planen",
   "getDirectionsForLocation": "Wegbeschreibung für {{address}} abrufen",
