+++ packages/visual-editor/locales/platform/en-GB/visual-editor.json
@@ -523,7 +523,6 @@
     "search": "Search fonts..."
   },
   "fontSize": "Font Size",
-  "fontSizeDefaultLabel": "Default",
   "friday": "Friday",
   "getDirections": "Get Directions",
   "getDirectionsForLocation": "Get Directions for {{address}}",
