+++ packages/visual-editor/locales/platform/es/visual-editor.json
@@ -523,7 +523,6 @@
     "search": "Fuentes de búsqueda ..."
   },
   "fontSize": "Tamaño de fuente",
-  "fontSizeDefaultLabel": "Por defecto",
   "friday": "Viernes",
   "getDirections": "Obtener instrucciones",
   "getDirectionsForLocation": "Obtener indicaciones para {{address}}",
