+++ packages/visual-editor/locales/platform/et/visual-editor.json
@@ -523,7 +523,6 @@
     "search": "Otsingufondid ..."
   },
   "fontSize": "Fondi suurus",
-  "fontSizeDefaultLabel": "Täitmata jätmine",
   "friday": "Reede",
   "getDirections": "Hankige juhiseid",
   "getDirectionsForLocation": "Hankige juhised {{address}}",
