+++ packages/visual-editor/locales/platform/fi/visual-editor.json
@@ -523,7 +523,6 @@
     "search": "Haku fontit ..."
   },
   "fontSize": "Fontin koko",
-  "fontSizeDefaultLabel": "Laiminlyönti",
   "friday": "perjantai",
   "getDirections": "Hanki reittiohjeet",
   "getDirectionsForLocation": "Hanki reittiohjeet kohteeseen {{address}}",
