+++ packages/visual-editor/locales/platform/hr/visual-editor.json
@@ -523,7 +523,6 @@
     "search": "Pretražite fontove ..."
   },
   "fontSize": "Veličina fonta",
-  "fontSizeDefaultLabel": "Zadano",
   "friday": "petak",
   "getDirections": "Dobiti upute",
   "getDirectionsForLocation": "Preuzmite upute za {{address}}",
