+++ packages/visual-editor/locales/platform/hu/visual-editor.json
@@ -523,7 +523,6 @@
     "search": "Keresési betűtípusok ..."
   },
   "fontSize": "Betűkészlet",
-  "fontSizeDefaultLabel": "Mulasztás",
   "friday": "péntek",
   "getDirections": "Szerezzen útmutatásokat",
   "getDirectionsForLocation": "Útvonaltervezés ehhez: {{address}}",
