+++ packages/visual-editor/locales/platform/it/visual-editor.json
@@ -523,7 +523,6 @@
     "search": "Cerca caratteri ..."
   },
   "fontSize": "Dimensione del carattere",
-  "fontSizeDefaultLabel": "Predefinito",
   "friday": "Venerdì",
   "getDirections": "Ottieni indicazioni",
   "getDirectionsForLocation": "Ottieni indicazioni per {{address}}",
