+++ packages/visual-editor/locales/platform/lt/visual-editor.json
@@ -523,7 +523,6 @@
     "search": "Ieškokite šriftų ..."
   },
   "fontSize": "Šrifto dydis",
-  "fontSizeDefaultLabel": "Numatytasis",
   "friday": "Penktadienis",
   "getDirections": "Gaukite nurodymus",
   "getDirectionsForLocation": "Gaukite nuorodas į {{address}}",
