+++ packages/visual-editor/locales/platform/nb/visual-editor.json
@@ -523,7 +523,6 @@
     "search": "Søkefonter ..."
   },
   "fontSize": "Fontstørrelse",
-  "fontSizeDefaultLabel": "Misligholde",
   "friday": "fredag",
   "getDirections": "Få veibeskrivelse",
   "getDirectionsForLocation": "Få veibeskrivelse for {{address}}",
