+++ packages/visual-editor/locales/platform/nl/visual-editor.json
@@ -523,7 +523,6 @@
     "search": "Zoek lettertypen ..."
   },
   "fontSize": "Lettergrootte",
-  "fontSizeDefaultLabel": "Standaard",
   "friday": "Vrijdag",
   "getDirections": "Routebeschrijving",
   "getDirectionsForLocation": "Routebeschrijving opvragen voor {{address}}",
