+++ packages/visual-editor/locales/platform/pl/visual-editor.json
@@ -523,7 +523,6 @@
     "search": "Wyszukaj czcionki ..."
   },
   "fontSize": "Rozmiar czcionki",
-  "fontSizeDefaultLabel": "Domyślny",
   "friday": "Piątek",
   "getDirections": "Uzyskaj wskazówki",
   "getDirectionsForLocation": "Uzyskaj wskazówki dojazdu do {{address}}",
