+++ packages/visual-editor/locales/platform/pt/visual-editor.json
@@ -523,7 +523,6 @@
     "search": "Pesquisar fontes ..."
   },
   "fontSize": "Tamanho da fonte",
-  "fontSizeDefaultLabel": "Padrão",
   "friday": "Sexta-feira",
   "getDirections": "Obtenha direções",
   "getDirectionsForLocation": "Obtenha rotas para {{address}}",
