+++ packages/visual-editor/locales/platform/ro/visual-editor.json
@@ -523,7 +523,6 @@
     "search": "Căutați fonturi ..."
   },
   "fontSize": "Dimensiunea fontului",
-  "fontSizeDefaultLabel": "Implicit",
   "friday": "vineri",
   "getDirections": "Obțineți indicații",
   "getDirectionsForLocation": "Obțineți indicații pentru {{address}}",
