+++ packages/visual-editor/locales/platform/sk/visual-editor.json
@@ -523,7 +523,6 @@
     "search": "Vyhľadajte písma ..."
   },
   "fontSize": "Veľkosť písma",
-  "fontSizeDefaultLabel": "Predvolený",
   "friday": "Piatok",
   "getDirections": "Získajte pokyny",
   "getDirectionsForLocation": "Získať trasu pre {{address}}",
