+++ packages/visual-editor/src/fields/index.ts
@@ -37,18 +37,6 @@ export {
 } from "./OptionalNumberField.tsx";
 
 export {
-  StyledButtonFieldOverride,
-  type StyledButtonField,
-  type StyledButtonValue,
-} from "./StyledButtonField.tsx";
-
-export {
-  StyledTextFieldOverride,
-  type StyledTextField,
-  type StyledTextValue,
-} from "./StyledTextField.tsx";
-
-export {
   TranslatableStringFieldOverride,
   type TranslatableStringField,
 } from "./TranslatableStringField.tsx";
