+++ packages/visual-editor/src/fields/styledFields/index.ts
@@ -1,0 +1,11 @@
+export {
+  StyledTextFieldOverride,
+  type StyledTextField,
+  type StyledTextValue,
+} from "./StyledTextField.tsx";
+
+export {
+  StyledButtonFieldOverride,
+  type StyledButtonField,
+  type StyledButtonValue,
+} from "./StyledButtonField.tsx";
